Invoke libtoolize in build.sh.
The following sequence fails for me:
./build.sh ; ./configure ; make
In fact, the "make" process blows up with lots of errors from libtool.
The problem is that the ltmain.sh distributed in speechd-up's git repo
is out of date.  Instead, build.sh should call libtoolize.  This copies
ltmain.sh from the host system.

Note that ltmain.sh should be removed from the repository.  I'd do that
in this patch, but it would needlessly increase the size of the diff.