Insure that text is NUL-terminated before speaking it.
In parse_buf, if text preceded a command, speak was called to render
it.  However, the text was not NUL-terminated.
So suppose we get the string \x01+3pX\x01-3p
That renders the x, with a pitch increment preceding it and a pitch
decrement succeeding it.  But the string X is not NUL-terminated.
Have you heard any odd characters when you try to read capital letters,
for example?  This is the cause.