change the way ssml processing is handled
We should escape each character of text as it comes in.
Previously, we were escaping the whole buffer, before it was spoken.
That was from my patch, previously sent to the Speech Dispatcher list.
The problem is that index marks are escaped, along with the text,
and this breaks the read-all feature.